lab_1/requirements.txt
ultralytics
PyYAML
Pillow

